requirements.txt
fastapi==0.133.1
httpx==0.28.1
python-dotenv==1.2.1
uvicorn==0.41.0
